# swagger/swagger.json
{"swagger":"2.0","info":{"title":"An api to handle appointments","version":""},"host":"localhost:9000","schemes":["http"],"consumes":["application/json","application/xml","application/gob","application/x-gob"],"produces":["application/json","application/xml","application/gob","application/x-gob"],"paths":{"/appointments":{"get":{"tags":["calendar"],"summary":"list calendar","description":"show a list of appointments for the selected time. If no interval is selected it shows the appointments of the current week\n\nRequired security scopes:\n","operationId":"calendar#list","parameters":[{"name":"end","in":"query","description":"The end of the interval","required":false,"type":"string"},{"name":"start","in":"query","description":"The start of the interval","required":false,"type":"string"},{"name":"where","in":"query","description":"Restrict the query to a location","required":false,"type":"string","default":""}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/MaraAppointmentCollection"}},"500":{"description":"Internal Server Error"}},"schemes":["http"],"security":[{"jwt":[]}]},"put":{"tags":["calendar"],"summary":"create calendar","description":"create a new appointment\n\nRequired security scopes:\n","operationId":"calendar#create","parameters":[{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/Appointment"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/MaraAppointment"}},"400":{"description":"Bad Request"},"500":{"description":"Internal Server Error"}},"schemes":["http"],"security":[{"jwt":[]}]}},"/appointments/{id}":{"get":{"tags":["calendar"],"summary":"show calendar","description":"show the details of a single appointment\n\nRequired security scopes:\n","operationId":"calendar#show","parameters":[{"name":"id","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/MaraAppointment"}},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}},"schemes":["http"],"security":[{"jwt":[]}]},"post":{"tags":["calendar"],"summary":"edit calendar","description":"change properties of a single appointment\n\nRequired security scopes:\n","operationId":"calendar#edit","parameters":[{"name":"id","in":"path","required":true,"type":"string"},{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/Appointment"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/MaraAppointment"}},"400":{"description":"Bad Request"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}},"schemes":["http"],"security":[{"jwt":[]}]},"delete":{"tags":["calendar"],"summary":"delete calendar","description":"deletes a single appointment\n\nRequired security scopes:\n","operationId":"calendar#delete","parameters":[{"name":"id","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/MaraAppointment"}},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}},"schemes":["http"],"security":[{"jwt":[]}]}},"/auth/login":{"post":{"tags":["auth"],"summary":"login auth","description":"Login with username and password and obtain a token","operationId":"auth#login","parameters":[{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/Login"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/MaraToken"}},"400":{"description":"Bad Request"}},"schemes":["http"]}},"/auth/reset":{"post":{"tags":["auth"],"summary":"reset auth","description":"Send a new password to the email specified","operationId":"auth#reset","parameters":[{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/Reset"}}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request"}},"schemes":["http"]}},"/builder/v1/swagger.json":{"get":{"summary":"Download swagger/swagger.json","operationId":"public#/builder/v1/swagger.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["http"]}}},"definitions":{"Appointment":{"title":"Appointment","type":"object","properties":{"email":{"type":"string","description":"The email of who's requesting it","example":"user@example.com","format":"email"},"notes":{"type":"string","description":"Internal notes about the Appointment","example":"Repudiandae eveniet tempora."},"phone":{"type":"string","description":"The phone of the client","example":"3347662313"},"problematic":{"type":"boolean","description":"Option specifying that the work is problematic for some reason","default":false,"example":false},"send_email":{"type":"boolean","description":"Option to send email to the client","default":false,"example":false},"send_sms":{"type":"boolean","description":"Option to send sms to the client","default":false,"example":true},"status":{"type":"string","description":"The status of the work. Can be one of todo/doing/done","example":"todo","pattern":"todo|doing|done"},"urgent":{"type":"boolean","description":"Option specifying that the work is to be done now","default":false,"example":false},"what":{"type":"string","description":"What is the purpose of the appointment","example":"Convergenza"},"when":{"type":"string","description":"When is the appointment scheduled","example":"2004-04-25T13:13:35+02:00","format":"date-time"},"where":{"type":"string","description":"Where is the appointment scheduled","example":"Budrio"},"who":{"type":"string","description":"Who is requesting the appointment","example":"Pippo Pippis"}},"example":{"email":"user@example.com","notes":"Repudiandae eveniet tempora.","phone":"3347662313","problematic":false,"send_email":false,"send_sms":true,"status":"todo","urgent":false,"what":"Convergenza","when":"2004-04-25T13:13:35+02:00","where":"Budrio","who":"Pippo Pippis"},"required":["who","what","when","where"]},"Login":{"title":"Login","type":"object","properties":{"password":{"type":"string","description":"The password","example":"password"},"user":{"type":"string","description":"The username","example":"user"}},"example":{"password":"password","user":"user"},"required":["user","password"]},"MaraAppointment":{"title":"Mediatype identifier: application/vnd.mara.appointment","type":"object","properties":{"email":{"type":"string","description":"The email of who's requesting it","example":"user@example.com","format":"email"},"href":{"type":"string","description":"The url of the appointment","example":"Eligendi natus."},"id":{"type":"string","description":"The id of the appointment","example":"Dicta consectetur illo mollitia qui qui pariatur."},"notes":{"type":"string","description":"Internal notes about the Appointment","example":"Doloremque suscipit aut sed omnis fugit accusamus."},"phone":{"type":"string","description":"The phone of the client","example":"3347662313"},"problematic":{"type":"boolean","description":"Option specifying that the work is problematic for some reason","default":false,"example":true},"send_email":{"type":"boolean","description":"Option to send email to the client","default":false,"example":true},"send_sms":{"type":"boolean","description":"Option to send sms to the client","default":false,"example":true},"status":{"type":"string","description":"The status of the work. Can be one of todo/doing/done","example":"todo","pattern":"todo|doing|done"},"urgent":{"type":"boolean","description":"Option specifying that the work is to be done now","default":false,"example":true},"what":{"type":"string","description":"What is the purpose of the appointment","example":"Convergenza"},"when":{"type":"string","description":"When is the appointment scheduled","example":"1988-01-03T19:41:46+01:00","format":"date-time"},"where":{"type":"string","description":"Where is the appointment scheduled","example":"Budrio"},"who":{"type":"string","description":"Who is requesting the appointment","example":"Pippo Pippis"}},"description":"MaraAppointment media type (default view)","example":{"email":"user@example.com","href":"Eligendi natus.","id":"Dicta consectetur illo mollitia qui qui pariatur.","notes":"Doloremque suscipit aut sed omnis fugit accusamus.","phone":"3347662313","problematic":true,"send_email":true,"send_sms":true,"status":"todo","urgent":true,"what":"Convergenza","when":"1988-01-03T19:41:46+01:00","where":"Budrio","who":"Pippo Pippis"}},"MaraAppointmentCollection":{"title":"Mediatype identifier: application/vnd.mara.appointment; type=collection","type":"array","items":{"$ref":"#/definitions/MaraAppointment"},"description":"MaraAppointmentCollection is the media type for an array of MaraAppointment (default view)","example":[{"email":"user@example.com","href":"Eligendi natus.","id":"Dicta consectetur illo mollitia qui qui pariatur.","notes":"Doloremque suscipit aut sed omnis fugit accusamus.","phone":"3347662313","problematic":true,"send_email":true,"send_sms":true,"status":"todo","urgent":true,"what":"Convergenza","when":"1988-01-03T19:41:46+01:00","where":"Budrio","who":"Pippo Pippis"},{"email":"user@example.com","href":"Eligendi natus.","id":"Dicta consectetur illo mollitia qui qui pariatur.","notes":"Doloremque suscipit aut sed omnis fugit accusamus.","phone":"3347662313","problematic":true,"send_email":true,"send_sms":true,"status":"todo","urgent":true,"what":"Convergenza","when":"1988-01-03T19:41:46+01:00","where":"Budrio","who":"Pippo Pippis"}]},"MaraToken":{"title":"Mediatype identifier: application/vnd.mara.token","type":"object","properties":{"token":{"type":"string","description":"The token to use in subsequent api calls","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ"}},"description":"MaraToken media type (default view)","example":{"token":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ"}},"Reset":{"title":"Reset","type":"object","properties":{"user":{"type":"string","description":"The user that will receive the new password","example":"user"}},"example":{"user":"user"},"required":["user"]}},"responses":{"BadRequest":{"description":"Bad Request"},"InternalServerError":{"description":"Internal Server Error"},"NotFound":{"description":"Not Found"},"OK":{"description":"OK"}},"securityDefinitions":{"jwt":{"type":"apiKey","name":"Authorization","in":"header"}}}